SW: fix sector object setup.
When moving the SO out of the way to get the sector, this only works for rotatable SO's. For non-rotating ones the internally used functions never did what was expected, letting the SO end up in invalid space.
To do this properly the SOBJ_DONT_ROTATE flag must be cleared temporarily.

+++ source/games/sw/src/track.cpp
@@ -1766,7 +1766,7 @@ PlayerPart:
 
 void RefreshPoints(SECTOR_OBJECT* sop, const DVector2& move, bool dynamic)
 {
-    short wallcount = 0;
+    int wallcount = 0;
     DAngle delta_ang_from_orig;
 
     // do scaling
